+++ resources/js/types/Comment.ts
@@ -1,0 +1,9 @@
+export type Comment = {
+    author_id: number
+    username: string
+    content: string
+    created_at: string
+    id: number
+    post_id: number
+    updated_at: string
+}
