+++ resources/js/types/Post.ts
@@ -19,12 +19,3 @@ export type NewPost = {
     longitude:number
     color:string
 }
-
-export type Comment = {
-    author_id: number
-    content: string
-    created_at: string
-    id: number
-    post_id: number
-    updated_at: string
-}
